Added extension method for validation
